apps/ws-server/pyproject.toml
[project]
name = "ws-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.14"
dependencies = [
    "aiortc>=1.14.0",
    "fastapi>=0.128.0",
    "faster-whisper>=1.2.1",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "scipy>=1.17.0",
    "setuptools>=80.9.0",
    "uvicorn>=0.40.0",
    "webrtcvad>=2.0.10",
    "websockets>=16.0",
]
